Issue #886: Consolidate safeEqual util
