+++ modules/security/index.ts
@@ -1,3 +1,4 @@
 // Security module public surface.
 export * from './internal/redaction.js';
+export * from './internal/safe-equal.js';
 export * from './internal/signed-ws-token.js';
